+++ py3arch/__init__.py
@@ -1,2 +1,2 @@
 from .core_modules import list_core_modules  # noqa: F401
-from .collect import collect_modules, find_imports  # noqa: F401
+from .collect import collect_modules, extract_imports_ast  # noqa: F401
